+++ docs/source/changelog.rst
@@ -4,6 +4,8 @@ Changelog
 
 * Now SocketServer starts correspondant listener on 'l' command
 * **BREAKING** DataProcessingChannel.process_queue and AbstractChannel.message_stream now cannot be awaited
+* **BREAKING** Changed DataProcessingChannel.process_queue 'l' parameter changed to 'listener'
+
 
 
 0.7.6